pyproject.toml
[project]
name = "arena4-capstone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "nnsight>=0.3.3",
    # Pinned to numpy<2 due to compatibility issues with transformers library
    # TODO: Remove this constraint once transformers adds numpy 2.0 support
    "numpy<2.0.0",
    "openai>=1.52.2",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "torch<2.3.0",
    "tqdm>=4.66.5",
    "transformers>=4.46.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
